biler svinger i forskjellige retninger, men kresjer i av og til i krysset. biler tar også å steller seg oppå hverandre

+++ src/main/java/com/oblig/obj_oblig_2/TrafficLight.java
@@ -70,8 +70,8 @@ public class TrafficLight {
         }
     
         gc.setFill(lightColor);
-        gc.fillOval(position.getX() - SIZE / 2, position.getY() - SIZE / 2, SIZE, SIZE);
+        gc.fillOval(position.getX() - (double) SIZE / 2, position.getY() - (double) SIZE / 2, SIZE, SIZE);
         gc.setStroke(Color.BLACK);
-        gc.strokeOval(position.getX() - SIZE / 2, position.getY() - SIZE / 2, SIZE, SIZE);
+        gc.strokeOval(position.getX() - (double) SIZE / 2, position.getY() - (double) SIZE / 2, SIZE, SIZE);
     }
 }
